search live all table

+++ app/Http/Controllers/DivisionController.php
@@ -16,13 +16,49 @@ class DivisionController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
         $divisi = Division::paginate(5);
         foreach ($divisi as $item) {
             $jumlah_posisi = Position::where('division_id', $item->id)->count();
             $item->jumlah_posisi = $jumlah_posisi;
         }
+        
+        if ($request->ajax()) {
+            $query = $request->input('query');
+            $divisi = Division::where('name', 'LIKE', '%' . $query . '%')->paginate(5);
+            foreach ($divisi as $item) {
+                $jumlah_posisi = Position::where('division_id', $item->id)->count();
+                $item->jumlah_posisi = $jumlah_posisi;
+            }
+
+
+        $output = '';
+        $iteration = 0;
+
+        foreach ($divisi as $item) {
+        $iteration++;
+            $output .= '<tr class="intro-x h-16">
+                <td class="w-4 text-center">' . $iteration . '.</td>
+                <td class="w-50 text-center capitalize">' . $item->name . '</td>
+                <td class="w-50 text-center capitalize">' . $item->jumlah_posisi . ' Posisi</td>
+                <td class="table-report__action w-56">
+                    <div class="flex justify-center items-center">
+                    <a class="flex items-center text-warning mr-3 edit-modal-divisi-search-class" data-Divisiid="'. $item->id .'" data-DivisiName="'. $item->name .'" href="javascript:;" data-tw-toggle="modal" data-tw-target="#modal-edit-divisi-search">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" icon-name="check-square" data-lucide="check-square" class="lucide lucide-check-square w-4 h-4 mr-1"><polyline points="9 11 12 14 22 4"></polyline><path d="M21 12v7a2 2 0 01-2 2H5a2 2 0 01-2-2V5a2 2 0 012-2h11"></path></svg> Edit
+                        </a>
+                        <a class="flex items-center text-danger delete-divisi-modal-search" data-DeleteDivisiId="'. $item->id .'" data-DeleteDivisiName="'. $item->name .'" href="javascript:;" data-tw-toggle="modal" data-tw-target="#delete-confirmation-modal-search">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" icon-name="check-square" data-lucide="check-square" class="lucide lucide-check-square w-4 h-4 mr-1"><polyline points="9 11 12 14 22 4"></polyline><path d="M21 12v7a2 2 0 01-2 2H5a2 2 0 01-2-2V5a2 2 0 012-2h11"></path></svg> Delete
+                        </a>
+                    </div>
+                </td>
+            </tr>';
+        }
+
+        return response($output);
+
+        }
+
         return view('divisi.index',compact('divisi'));
     }
 
@@ -88,11 +124,12 @@ class DivisionController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(string $id, Request $request)
     {
         try {
             $divisi = Division::findOrFail($id);
             $posisi = Position::where('division_id', $divisi->id)->get();
+            $inputName= $request->input('validName');
 
             foreach ($posisi as $posisiItem) {
                 if ($posisiItem->employee->count() > 0) {
@@ -101,10 +138,14 @@ class DivisionController extends Controller
                 $posisiItem->delete();
             }
 
-            $nama_divisi = $divisi->division;
-            $divisi->delete();
+         if($inputName === $divisi->name){            
+             $nama_divisi = $divisi->division;
+             $divisi->delete();
+             return redirect()->back()->with(['delete' => "$nama_divisi deleted successfully"]);
+         }else {
+            return redirect()->back()->with(['error' => 'Wrong username']);
+         };
 
-            return redirect()->back()->with(['delete' => "$nama_divisi deleted successfully"]);
         } catch (\Throwable $th) {
             return redirect()->back()->with(['error' => "Failed to delete division"]);
         }
